+++ wordpool/listgen/__init__.py
@@ -1,31 +1,46 @@
 """List generation and I/O."""
 
-from __future__ import print_function
+try:
+    import random
+except:
+    random = None
+try:
+    import os.path as osp
+except:
+    osp = None
+try:
+    import itertools
+except:
+    itertools = None
+try:
+    import numpy.random as npr
+except:
+    npr = None
+try:
+    import pandas as pd
+except:
+    pd = None
 
-import random
-import os.path as osp
-import itertools
+try:
+    from .. import load, pool_dataframe_to_pool_list, pool_list_to_pool_dataframe, exc
+except ImportError:
+    load = pool_dataframe_to_pool_list = pool_list_to_pool_dataframe = exc = None
+try:
+    from . import fr, catfr, pal
+except ImportError:
+    fr = catfr = pal = None
 
-import numpy.random as npr
-import pandas as pd
+if (load):
+    RAM_LIST_EN = load("ram_wordpool_en.txt")
+    RAM_LIST_SP = load("ram_wordpool_sp.txt")
 
-from .. import load, pool_dataframe_to_pool_list, pool_list_to_pool_dataframe
-from .. import exc
-from . import fr
-from . import catfr
-from . import pal
-from init_no_pandas import *
+    CAT_LIST_EN = load("ram_categorized_en.txt")
+    CAT_LIST_SP = load("ram_categorized_sp.txt")
 
-RAM_LIST_EN = load("ram_wordpool_en.txt")
-RAM_LIST_SP = load("ram_wordpool_sp.txt")
+    PRACTICE_LIST_EN = load("practice_en.txt")
+    PRACTICE_LIST_SP = load("practice_sp.txt")
 
-CAT_LIST_EN = load("ram_categorized_en.txt")
-CAT_LIST_SP = load("ram_categorized_sp.txt")
-
-PRACTICE_LIST_EN = load("practice_en.txt")
-PRACTICE_LIST_SP = load("practice_sp.txt")
-
-LURES_LIST_EN = load("REC1_lures_en.txt")
+    LURES_LIST_EN = load("REC1_lures_en.txt")
 
 
 def write_wordpool_txt(path, language="EN", include_lure_words=False,
@@ -100,10 +115,53 @@ def assign_list_types(pool, num_baseline, num_nonstim, num_stim, num_ps=0):
     
     
     pool_list = pool_dataframe_to_pool_list(pool)
-    pool_list = assign_list_types_no_pandas(pool_list, num_baseline, stim_or_nostim, num_ps = num_ps)
+    pool_list = assign_list_types_from_type_list(pool_list, num_baseline, stim_or_nostim, num_ps = num_ps)
     pool_dataframe = pool_list_to_pool_dataframe(pool_list)
     
     return pool_dataframe
+
+def assign_list_types_from_type_list(pool, num_baseline, stim_nonstim, num_ps=0):
+    """Assign list types to a pool. The types are:
+        
+        * ``PRACTICE``
+        * ``BASELINE``
+        * ``PS``
+        * ``STIM``
+        * ``NON-STIM``
+        
+        :param pd.DataFrame pool: Input word pool.  list of (word, listno) pairs
+        :param int num_baseline: Number of baseline trials *excluding* the practice
+        list.
+        :param list stim_nonstim: a list of "STIM" or "NON-STIM" strings indicating the order of stim and non-stim interleaved lists.
+        :param int num_ps: Number of parameter search trials.
+        :returns: pool with assigned types
+        :rtype: list
+        
+        """
+    
+    # Check that the inputs match the number of lists
+    last_listno = pool[-1]['listno']
+    assert last_listno == num_baseline + len(stim_nonstim) + num_ps, "The number of lists and provided type parameters didn't match"
+    
+    
+    for i in range(len(pool)):
+        word = pool[i]
+        if (word['listno'] == 0):
+            pool[i]['type'] = "PRACTICE"
+            pool[i]['stim_channels'] = None
+        elif (word['listno'] <= num_baseline):
+            pool[i]['type'] = "BASELINE"
+            pool[i]['stim_channels'] = None
+        elif (word['listno'] <= num_baseline + num_ps):
+            pool[i]['type'] = "PS"
+            pool[i]['stim_channels'] = None
+        else:
+            stimtype = stim_nonstim[word['listno']-num_ps-num_baseline-1]
+            pool[i]['type'] = stimtype
+            pool[i]['stim_channels'] = (0,) if stimtype == "STIM" else None
+
+    return pool
+
 
 def assign_multistim(pool, stimspec):
     """Update stim lists to account for multiple stimulation sites.
@@ -139,10 +197,43 @@ def assign_multistim(pool, stimspec):
     random.shuffle(stimspec_list)
     
     pool_list = pool_dataframe_to_pool_list(pool)
-    pool_list = assign_multistim_no_pandas(pool_list, stimspec_list)
+    pool_list = assign_multistim_from_stim_channels_list(pool_list, stimspec_list)
     pool_dataframe = pool_list_to_pool_dataframe(pool_list)
     
     return pool_dataframe
+
+def assign_multistim_from_stim_channels_list(pool, stimspec_list):
+    """Update stim lists to account for multiple stimulation sites.
+        
+        
+        :param list pool: Word pool with assigned stim lists. (word, listno, stim_channels, type)
+        :param list names: Names of individual stim channels.
+        :rtype: list
+        
+        """
+    assert len(pool) > 0, "Empty pool"
+    assert len(pool[0]) == 4, "Pool should be a list of four-tuples"
+    
+    stim_words = [word for word in pool if word['type'] == "STIM"]
+    unique_listnos = set()
+    for word in stim_words:
+        unique_listnos.add(word['listno'])
+
+    assert len(unique_listnos) == len(stimspec_list), "The number of stimspecs should be the same as the number of stim lists."
+
+    current_stimspec_index = -1
+    stim_listno = -1
+    for i in range(len(pool)):
+        word = pool[i]
+        if (word['type'] == "STIM"):
+            if (word['listno'] != stim_listno):
+                stim_listno = word['listno']
+                current_stimspec_index += 1
+            pool[i]['stim_channels'] = stimspec_list[current_stimspec_index]
+
+
+    return pool
+
 
 
 
@@ -215,3 +306,31 @@ def generate_learn1_blocks(pool, num_nonstim, num_stim, stim_channels=(0,1), num
     result = pool_list_to_pool_dataframe(result_list)
 
     return result
+
+
+
+def extract_blocks(pool, listnos, num_blocks):
+    """Take out lists based on listnos and separate them into blocks
+        
+        :param list pool: Input word pool.
+        :param list listnos: The order of lists to separate into blocks
+        :param int num_blocks: The number of blocks to organize the listnos into
+        :returns: blocks of words as a list of tuples
+        
+        """
+    assert len(listnos)%num_blocks == 0, "The number of lists to append must be divisable by the number of blocks"
+    
+    wordlists = {}
+    for word in pool:
+        wordlists[word['listno']] = wordlists.get(word['listno'], []) + [word]
+    
+    blocks = []
+    for i in range(len(listnos)):
+        listno = listnos[i]
+        wordlist = [word.copy() for word in wordlists[listno]]
+        for word in wordlist:
+            word['blockno'] = i/num_blocks
+            word['block_listno'] = i
+        blocks += wordlist
+    
+    return blocks
